Code editor files simplification refactor
